fix windows reboot

+++ src/command/reboot.rs
@@ -1,4 +1,4 @@
-use std::process;
+use std::{process, env};
 
 use async_trait::async_trait;
 use tracing::info;
@@ -14,11 +14,14 @@ impl Command for RebootCommand {
     async fn execute(&self, mut _params: Vec<String>) -> eyre::Result<CommandResp> {
         info!("executing reboot command");
 
+        if env::consts::OS == "windows" {
+            info!("reboot command not available for Windows OS");
+            return Ok(CommandResp { status: Status::NotFound });
+        }
+
         let output = process::Command::new("reboot").output()?;
         info!("reboot command executed with exit code {}", output.status);
 
-        Ok(CommandResp {
-            status: Status::Success,
-        })
+        Ok(CommandResp { status: Status::Success, })
     }
 }
